langgraph/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
]
